moved engine trace
