+++ src/modules/keyboardmanager/KeyboardManagerEngineLibrary/KeyboardManager.cpp
@@ -3,16 +3,18 @@
 #include <interface/powertoy_module_interface.h>
 #include <common/SettingsAPI/settings_objects.h>
 #include <common/interop/shared_constants.h>
-#include <keyboardmanager/common/Shortcut.h>
-#include <keyboardmanager/common/RemapShortcut.h>
-#include <keyboardmanager/common/KeyboardManagerConstants.h>
 #include <common/debug_control.h>
 #include <common/utils/winapi_error.h>
 #include <common/logger/logger_settings.h>
-#include <keyboardmanager/common/trace.h>
+
+#include <keyboardmanager/common/Shortcut.h>
+#include <keyboardmanager/common/RemapShortcut.h>
+#include <keyboardmanager/common/KeyboardManagerConstants.h>
 #include <keyboardmanager/common/Helpers.h>
-#include "KeyboardEventHandlers.h"
 #include <keyboardmanager/common/KeyboardEventHandlers.h>
+
+#include "KeyboardEventHandlers.h"
+#include "trace.h"
 
 HHOOK KeyboardManager::hook_handle_copy;
 HHOOK KeyboardManager::hook_handle;
